requirements.txt
faiss-cpu>=1.7.4
faiss-gpu>=1.7.4
numpy>=1.21.0
pydantic>=2.0.0
typing-extensions>=4.0.0
pynvml>=11.0.0
